refactor(map/aircraft): split AircraftCreationManager into manager + form
AircraftCreationManager.ts mixed the map-click flow (placing a new
aircraft on the map, capturing heading) with the modal form (fields,
validation, submit). The form is now its own module so the manager
can focus on the map interaction.

- AircraftCreationForm: owns the modal lifecycle, the type/category
  inputs, and submit validation
- AircraftCreationManager: thinned to the map placement state
  machine and delegating to the form when needed
